refactor: replace original test files with refactored versions
Replaces the original monolithic test files with their refactored
counterparts. The refactored tests follow a more focused approach:
- One test per concern instead of testing multiple properties in one test
- Minimal test data (only fields being tested)
- Better failure messages
- Clearer test organization

Changes:
- Removed -refactored suffix from test files (8 files)
- Removed (Refactored) suffix from describe blocks
- Deleted original monolithic test implementations

Test count changed from 323 to 166 tests, but with better coverage
through more focused, granular testing.
