--- packages/dynamic-form-material/src/lib/fields/input/mat-input-refactored.spec.ts
@@ -1,275 +1,0 @@
-import { By } from '@angular/platform-browser';
-import { MaterialFormTestUtils } from '../../testing/material-test-utils';
-
-describe('MatInputFieldComponent (Refactored)', () => {
-  describe('Input Type Attribute (Unit)', () => {
-    it('should render email input type', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'field', props: { type: 'email' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      expect(input.nativeElement.getAttribute('type')).toBe('email');
-    });
-
-    it('should render text input type', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'field', props: { type: 'text' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      expect(input.nativeElement.getAttribute('type')).toBe('text');
-    });
-
-    it('should render password input type', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'field', props: { type: 'password' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      expect(input.nativeElement.getAttribute('type')).toBe('password');
-    });
-
-    it('should default to text type when not specified', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      expect(input.nativeElement.getAttribute('type')).toBe('text');
-    });
-  });
-
-  describe('Label Rendering (Unit)', () => {
-    it('should render label text', async () => {
-      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', label: 'Email Address' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const label = fixture.debugElement.query(By.css('mat-label'));
-      expect(label.nativeElement.textContent.trim()).toBe('Email Address');
-    });
-
-    it('should not render label when not provided', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const label = fixture.debugElement.query(By.css('mat-label'));
-      expect(label).toBeNull();
-    });
-  });
-
-  describe('Hint/Help Text (Unit)', () => {
-    it('should render hint text', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'field',
-          type: 'input',
-          props: { hint: 'We will never share your email' },
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const hint = fixture.debugElement.query(By.css('mat-hint'));
-      expect(hint.nativeElement.textContent.trim()).toBe('We will never share your email');
-    });
-
-    it('should not render hint when not provided', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const hint = fixture.debugElement.query(By.css('mat-hint'));
-      expect(hint).toBeNull();
-    });
-  });
-
-  describe('TabIndex Attribute (Unit)', () => {
-    it('should apply tabindex attribute', async () => {
-      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', tabIndex: 5 }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      expect(input.nativeElement.getAttribute('tabindex')).toBe('5');
-    });
-
-    it('should not set tabindex when not provided', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      // Material may set default tabindex, we just verify it's not our custom value
-      const tabindex = input.nativeElement.getAttribute('tabindex');
-      expect(tabindex).not.toBe('5'); // Our test value from previous test
-    });
-  });
-
-  describe('CSS Classes (Unit)', () => {
-    it('should apply custom className to wrapper', async () => {
-      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', className: 'email-input' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      // className applies to the host element (df-mat-input component)
-      const hostElement = fixture.debugElement.query(By.css('df-mat-input'));
-      expect(hostElement.nativeElement.className).toContain('email-input');
-    });
-  });
-
-  describe('Material Appearance (Unit)', () => {
-    it('should render outline appearance', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'field', props: { appearance: 'outline' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const formField = fixture.debugElement.query(By.css('mat-form-field'));
-      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
-    });
-
-    it('should render fill appearance', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'field', props: { appearance: 'fill' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const formField = fixture.debugElement.query(By.css('mat-form-field'));
-      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-fill');
-    });
-
-    it('should default to outline appearance', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const formField = fixture.debugElement.query(By.css('mat-form-field'));
-      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
-    });
-  });
-
-  describe('Required State (Unit)', () => {
-    it('should mark input as required', async () => {
-      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', required: true }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      expect(input.nativeElement.hasAttribute('required')).toBe(true);
-    });
-
-    it('should not mark input as required by default', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      expect(input.nativeElement.hasAttribute('required')).toBe(false);
-    });
-  });
-
-  describe('Full Configuration (Integration)', () => {
-    /**
-     * This integration test verifies multiple properties work together.
-     * It's kept as ONE test because it tests the INTEGRATION of features,
-     * not individual features. But notice it still uses minimal initialValue.
-     */
-    it('should render input with all properties configured correctly', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'email',
-          type: 'input',
-          label: 'Email Address',
-          required: true,
-          tabIndex: 1,
-          className: 'email-input',
-          props: {
-            placeholder: 'Enter your email',
-            hint: 'We will never share your email',
-            type: 'email',
-            appearance: 'outline',
-          },
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { email: '' },
-      });
-
-      // Verify all features work together
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      const formField = fixture.debugElement.query(By.css('mat-form-field'));
-      const label = fixture.debugElement.query(By.css('mat-label'));
-      const hint = fixture.debugElement.query(By.css('mat-hint'));
-      const hostElement = fixture.debugElement.query(By.css('df-mat-input'));
-
-      expect(input.nativeElement.getAttribute('type')).toBe('email');
-      expect(input.nativeElement.getAttribute('tabindex')).toBe('1');
-      expect(input.nativeElement.hasAttribute('required')).toBe(true);
-      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
-      expect(label.nativeElement.textContent.trim()).toBe('Email Address');
-      expect(hint.nativeElement.textContent.trim()).toBe('We will never share your email');
-      expect(hostElement.nativeElement.className).toContain('email-input');
-    });
-  });
-});
