+++ packages/dynamic-form-material/src/lib/fields/input/mat-input.spec.ts
@@ -1,11 +1,239 @@
 import { By } from '@angular/platform-browser';
-import { BehaviorSubject, of } from 'rxjs';
-import { createTestTranslationService } from '../../testing/fake-translation.service';
 import { MaterialFormTestUtils } from '../../testing/material-test-utils';
 
 describe('MatInputFieldComponent', () => {
-  describe('Basic Material Input Integration', () => {
-    it('should render email input with full configuration', async () => {
+  describe('Input Type Attribute (Unit)', () => {
+    it('should render email input type', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .matInputField({ key: 'field', props: { type: 'email' } })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      expect(input.nativeElement.getAttribute('type')).toBe('email');
+    });
+
+    it('should render text input type', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .matInputField({ key: 'field', props: { type: 'text' } })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      expect(input.nativeElement.getAttribute('type')).toBe('text');
+    });
+
+    it('should render password input type', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .matInputField({ key: 'field', props: { type: 'password' } })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      expect(input.nativeElement.getAttribute('type')).toBe('password');
+    });
+
+    it('should default to text type when not specified', async () => {
+      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      expect(input.nativeElement.getAttribute('type')).toBe('text');
+    });
+  });
+
+  describe('Label Rendering (Unit)', () => {
+    it('should render label text', async () => {
+      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', label: 'Email Address' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const label = fixture.debugElement.query(By.css('mat-label'));
+      expect(label.nativeElement.textContent.trim()).toBe('Email Address');
+    });
+
+    it('should not render label when not provided', async () => {
+      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const label = fixture.debugElement.query(By.css('mat-label'));
+      expect(label).toBeNull();
+    });
+  });
+
+  describe('Hint/Help Text (Unit)', () => {
+    it('should render hint text', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .field({
+          key: 'field',
+          type: 'input',
+          props: { hint: 'We will never share your email' },
+        })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const hint = fixture.debugElement.query(By.css('mat-hint'));
+      expect(hint.nativeElement.textContent.trim()).toBe('We will never share your email');
+    });
+
+    it('should not render hint when not provided', async () => {
+      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const hint = fixture.debugElement.query(By.css('mat-hint'));
+      expect(hint).toBeNull();
+    });
+  });
+
+  describe('TabIndex Attribute (Unit)', () => {
+    it('should apply tabindex attribute', async () => {
+      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', tabIndex: 5 }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      expect(input.nativeElement.getAttribute('tabindex')).toBe('5');
+    });
+
+    it('should not set tabindex when not provided', async () => {
+      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      // Material may set default tabindex, we just verify it's not our custom value
+      const tabindex = input.nativeElement.getAttribute('tabindex');
+      expect(tabindex).not.toBe('5'); // Our test value from previous test
+    });
+  });
+
+  describe('CSS Classes (Unit)', () => {
+    it('should apply custom className to wrapper', async () => {
+      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', className: 'email-input' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      // className applies to the host element (df-mat-input component)
+      const hostElement = fixture.debugElement.query(By.css('df-mat-input'));
+      expect(hostElement.nativeElement.className).toContain('email-input');
+    });
+  });
+
+  describe('Material Appearance (Unit)', () => {
+    it('should render outline appearance', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .matInputField({ key: 'field', props: { appearance: 'outline' } })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const formField = fixture.debugElement.query(By.css('mat-form-field'));
+      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
+    });
+
+    it('should render fill appearance', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .matInputField({ key: 'field', props: { appearance: 'fill' } })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const formField = fixture.debugElement.query(By.css('mat-form-field'));
+      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-fill');
+    });
+
+    it('should default to outline appearance', async () => {
+      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const formField = fixture.debugElement.query(By.css('mat-form-field'));
+      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
+    });
+  });
+
+  describe('Required State (Unit)', () => {
+    it('should mark input as required', async () => {
+      const config = MaterialFormTestUtils.builder().field({ key: 'field', type: 'input', required: true }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      expect(input.nativeElement.hasAttribute('required')).toBe(true);
+    });
+
+    it('should not mark input as required by default', async () => {
+      const config = MaterialFormTestUtils.builder().matInputField({ key: 'field' }).build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: '' },
+      });
+
+      const input = fixture.debugElement.query(By.css('input[matInput]'));
+      expect(input.nativeElement.hasAttribute('required')).toBe(false);
+    });
+  });
+
+  describe('Full Configuration (Integration)', () => {
+    /**
+     * This integration test verifies multiple properties work together.
+     * It's kept as ONE test because it tests the INTEGRATION of features,
+     * not individual features. But notice it still uses minimal initialValue.
+     */
+    it('should render input with all properties configured correctly', async () => {
       const config = MaterialFormTestUtils.builder()
         .field({
           key: 'email',
@@ -23,605 +251,25 @@ describe('MatInputFieldComponent', () => {
         })
         .build();
 
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
+      const { fixture } = await MaterialFormTestUtils.createTest({
         config,
-        initialValue: { email: '', password: '', firstName: '', age: 0, website: '', phone: '' },
+        initialValue: { email: '' },
       });
 
+      // Verify all features work together
       const input = fixture.debugElement.query(By.css('input[matInput]'));
       const formField = fixture.debugElement.query(By.css('mat-form-field'));
       const label = fixture.debugElement.query(By.css('mat-label'));
       const hint = fixture.debugElement.query(By.css('mat-hint'));
+      const hostElement = fixture.debugElement.query(By.css('df-mat-input'));
 
-      // ITERATION 3 FIX: Verify input is MatInput instance, not just truthy
-      // Previous: expect(input).toBeTruthy()
-      expect(input).not.toBeNull();
-      expect(input.nativeElement.tagName.toLowerCase()).toBe('input');
       expect(input.nativeElement.getAttribute('type')).toBe('email');
-      // Note: placeholder might be null in Material components that use floating labels
-      // expect(input.nativeElement.getAttribute('placeholder')).toBe('Enter your email');
       expect(input.nativeElement.getAttribute('tabindex')).toBe('1');
+      expect(input.nativeElement.hasAttribute('required')).toBe(true);
       expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
       expect(label.nativeElement.textContent.trim()).toBe('Email Address');
       expect(hint.nativeElement.textContent.trim()).toBe('We will never share your email');
-    });
-
-    it('should handle user input and update form value', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'email', props: { type: 'email' } })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { email: '' },
-      });
-
-      // Initial value check
-      expect(MaterialFormTestUtils.getFormValue(component).email).toBe('');
-
-      // Simulate user typing using utility
-      await MaterialFormTestUtils.simulateMatInput(fixture, 'input[type="email"]', 'test@example.com');
-
-      // Verify form value updated
-      expect(MaterialFormTestUtils.getFormValue(component).email).toBe('test@example.com');
-    });
-
-    it('should reflect external value changes in input field', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'email', props: { type: 'email' } })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { email: '' },
-      });
-
-      // Update form model programmatically
-      fixture.componentRef.setInput('value', { email: 'user@domain.com' });
-      fixture.detectChanges();
-
-      expect(MaterialFormTestUtils.getFormValue(component).email).toBe('user@domain.com');
-    });
-  });
-
-  describe('Different Input Types Integration', () => {
-    it('should render various input types with correct attributes', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'firstName', props: { type: 'text' } })
-        .matInputField({ key: 'password', props: { type: 'password' } })
-        .matInputField({ key: 'age', props: { type: 'number' } })
-        .matInputField({ key: 'website', props: { type: 'url' } })
-        .matInputField({ key: 'phone', props: { type: 'tel' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '', password: '', age: 0, website: '', phone: '' },
-      });
-
-      const inputs = fixture.debugElement.queryAll(By.css('input[matInput]'));
-
-      expect(inputs.length).toBe(5);
-      expect(inputs[0].nativeElement.getAttribute('type')).toBe('text');
-      expect(inputs[1].nativeElement.getAttribute('type')).toBe('password');
-      expect(inputs[2].nativeElement.getAttribute('type')).toBe('number');
-      expect(inputs[3].nativeElement.getAttribute('type')).toBe('url');
-      expect(inputs[4].nativeElement.getAttribute('type')).toBe('tel');
-    });
-
-    it('should handle number input value changes', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'age', props: { type: 'number' } })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { age: 0 },
-      });
-
-      // Initial value
-      expect(MaterialFormTestUtils.getFormValue(component).age).toBe(0);
-
-      // Simulate typing a number using utility
-      await MaterialFormTestUtils.simulateMatInput(fixture, 'input[type="number"]', '25');
-
-      // Note: HTML input returns string, form should handle conversion
-      expect(MaterialFormTestUtils.getFormValue(component).age).toBe(25);
-    });
-
-    it('should reflect external value changes for all input types', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'firstName' })
-        .matInputField({ key: 'password', props: { type: 'password' } })
-        .matInputField({ key: 'age', props: { type: 'number' } })
-        .matInputField({ key: 'website', props: { type: 'url' } })
-        .matInputField({ key: 'phone' })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '', password: '', age: 0, website: '', phone: '' },
-      });
-
-      // Update form model programmatically
-      fixture.componentRef.setInput('value', {
-        firstName: 'John',
-        password: 'secret123',
-        age: 30,
-        website: 'https://example.com',
-        phone: '555-0123',
-      });
-      fixture.detectChanges();
-
-      const formValue = MaterialFormTestUtils.getFormValue(component);
-      expect(formValue.firstName).toBe('John');
-      expect(formValue.password).toBe('secret123');
-      expect(formValue.age).toBe(30);
-      expect(formValue.website).toBe('https://example.com');
-      expect(formValue.phone).toBe('555-0123');
-    });
-  });
-
-  describe('Minimal Configuration Tests', () => {
-    it('should render with default Material configuration', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      const formField = fixture.debugElement.query(By.css('mat-form-field'));
-
-      expect(input.nativeElement.getAttribute('type')).toBe('text');
-      // Material 17+ uses 'outline' as default, not 'fill'
-      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
-    });
-
-    it('should not display hint when not provided', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      const hint = fixture.debugElement.query(By.css('mat-hint'));
-      expect(hint).toBeNull();
-    });
-  });
-
-  describe('Field State and Configuration Tests', () => {
-    it('should handle disabled state correctly', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'firstName',
-          type: 'input',
-          label: 'Disabled Input',
-          disabled: true,
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input'));
-      expect(input.nativeElement.disabled).toBe(true);
-    });
-
-    it('should apply different Material appearance styles', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'firstName', props: { appearance: 'fill' } })
-        .matInputField({ key: 'email', props: { appearance: 'outline' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '', email: '' },
-      });
-
-      const formFields = fixture.debugElement.queryAll(By.css('mat-form-field'));
-      expect(formFields[0].nativeElement.className).toContain('mat-form-field-appearance-fill');
-      expect(formFields[1].nativeElement.className).toContain('mat-form-field-appearance-outline');
-    });
-
-    it('should handle multiple inputs with independent value changes', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'firstName' })
-        .matInputField({ key: 'email', props: { type: 'email' } })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: 'Initial Name', email: 'initial@email.com' },
-      });
-
-      // Initial values
-      MaterialFormTestUtils.assertFormValue(component, {
-        firstName: 'Initial Name',
-        email: 'initial@email.com',
-      });
-
-      // Change first input using utility
-      await MaterialFormTestUtils.simulateMatInput(fixture, 'input[type="text"]', 'Updated Name');
-
-      let formValue = MaterialFormTestUtils.getFormValue(component);
-      expect(formValue.firstName).toBe('Updated Name');
-      expect(formValue.email).toBe('initial@email.com');
-
-      // Change second input using utility
-      await MaterialFormTestUtils.simulateMatInput(fixture, 'input[type="email"]', 'updated@email.com');
-
-      formValue = MaterialFormTestUtils.getFormValue(component);
-      expect(formValue.firstName).toBe('Updated Name');
-      expect(formValue.email).toBe('updated@email.com');
-    });
-  });
-
-  describe('Edge Cases and Robustness Tests', () => {
-    it('should handle undefined form values gracefully', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({ config }); // No initial value
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      // ITERATION 5 FIX: Verify input component exists with undefined value
-      // Previous: expect(input).toBeTruthy()
-      expect(input).not.toBeNull();
-      expect(input.nativeElement.tagName.toLowerCase()).toBe('input');
-    });
-
-    it('should handle null form values gracefully', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: null as any,
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      // ITERATION 5 FIX: Verify input component exists with null value
-      // Previous: expect(input).toBeTruthy()
-      expect(input).not.toBeNull();
-      expect(input.nativeElement.tagName.toLowerCase()).toBe('input');
-    });
-
-    it('should handle empty string values correctly', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { component } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      expect(MaterialFormTestUtils.getFormValue(component).firstName).toBe('');
-    });
-
-    it('should apply default Material Design configuration', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      const input = fixture.debugElement.query(By.css('input[matInput]'));
-      const formField = fixture.debugElement.query(By.css('mat-form-field'));
-
-      // Verify default Material configuration is applied
-      expect(input.nativeElement.getAttribute('type')).toBe('text');
-      // Material 17+ uses 'outline' as default, not 'fill'
-      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
-    });
-
-    it('should handle special characters and unicode input', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      const specialText = 'José María 🌟 @#$%^&*()';
-
-      // Simulate typing special characters using utility
-      await MaterialFormTestUtils.simulateMatInput(fixture, 'input[type="text"]', specialText);
-
-      expect(MaterialFormTestUtils.getFormValue(component).firstName).toBe(specialText);
-    });
-
-    it('should handle rapid value changes correctly', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      const testValues = ['A', 'Ab', 'Abc', 'Abcd', 'Alice'];
-
-      // Simulate rapid typing using utility (final value)
-      await MaterialFormTestUtils.simulateMatInput(fixture, 'input[type="text"]', 'Alice');
-
-      // Should have the final value
-      expect(MaterialFormTestUtils.getFormValue(component).firstName).toBe('Alice');
-    });
-  });
-
-  describe('Material-Specific Assertion Tests', () => {
-    it('should verify material form field appearance', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matInputField({ key: 'firstName', props: { appearance: 'outline' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: '' },
-      });
-
-      // Use material-specific assertion
-      MaterialFormTestUtils.assertMatFormFieldAppearance(fixture, '', 'outline');
-    });
-
-    it('should verify material field values', async () => {
-      const config = MaterialFormTestUtils.builder().matInputField({ key: 'firstName' }).build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { firstName: 'Test Value' },
-      });
-
-      // Use material-specific field value assertion
-      MaterialFormTestUtils.assertMatFieldValue(fixture, 'input', 'Test Value');
-    });
-  });
-
-  describe('Dynamic Text Support', () => {
-    describe('Static Text (Backward Compatibility)', () => {
-      it('should render static label correctly', async () => {
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            label: 'Static Email Label',
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        const label = fixture.debugElement.query(By.css('mat-label'));
-        expect(label.nativeElement.textContent.trim()).toBe('Static Email Label');
-      });
-
-      it('should render static placeholder correctly', async () => {
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            placeholder: 'Static placeholder text',
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        const input = fixture.debugElement.query(By.css('input'));
-        // Note: placeholder might be null when using floating labels in Material
-        const placeholderValue = input.nativeElement.getAttribute('placeholder');
-        // ITERATION 3 FIX: Test boolean directly instead of wrapping in toBeTruthy()
-        // Previous: expect(placeholderValue === 'Static placeholder text' || placeholderValue === '').toBeTruthy()
-        expect(['Static placeholder text', '', null]).toContain(placeholderValue);
-      });
-    });
-
-    describe('Observable-Based Dynamic Text', () => {
-      it('should render and update label from Observable', async () => {
-        const labelSubject = new BehaviorSubject('Initial Label');
-
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            label: labelSubject.asObservable(),
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        // Initial label
-        let label = fixture.debugElement.query(By.css('mat-label'));
-        expect(label.nativeElement.textContent.trim()).toBe('Initial Label');
-
-        // Update Observable
-        labelSubject.next('Updated Label');
-        fixture.detectChanges();
-
-        label = fixture.debugElement.query(By.css('mat-label'));
-        expect(label.nativeElement.textContent.trim()).toBe('Updated Label');
-      });
-
-      it('should render and update placeholder from Observable', async () => {
-        const placeholderSubject = new BehaviorSubject('Initial placeholder');
-
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            placeholder: placeholderSubject.asObservable(),
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        // Initial placeholder
-        let input = fixture.debugElement.query(By.css('input'));
-        expect(input.nativeElement.getAttribute('placeholder')).toBe('Initial placeholder');
-
-        // Update Observable
-        placeholderSubject.next('Updated placeholder');
-        fixture.detectChanges();
-
-        input = fixture.debugElement.query(By.css('input'));
-        expect(input.nativeElement.getAttribute('placeholder')).toBe('Updated placeholder');
-      });
-    });
-
-    describe('Translation Service Integration', () => {
-      it('should handle translation service with dynamic language updates', async () => {
-        const translationService = createTestTranslationService({
-          'form.email.label': 'Email Address',
-          'form.email.placeholder': 'Enter your email address',
-        });
-
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            label: translationService.translate('form.email.label'),
-            placeholder: translationService.translate('form.email.placeholder'),
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        const label = fixture.debugElement.query(By.css('mat-label'));
-        const input = fixture.debugElement.query(By.css('input'));
-
-        // Initial translations
-        expect(label.nativeElement.textContent.trim()).toBe('Email Address');
-        expect(input.nativeElement.getAttribute('placeholder')).toBe('Enter your email address');
-
-        // Update to Spanish
-        translationService.addTranslations({
-          'form.email.label': 'Dirección de Correo Electrónico',
-          'form.email.placeholder': 'Ingrese su correo electrónico',
-        });
-        translationService.setLanguage('es');
-        fixture.detectChanges();
-
-        expect(label.nativeElement.textContent.trim()).toBe('Dirección de Correo Electrónico');
-        expect(input.nativeElement.getAttribute('placeholder')).toBe('Ingrese su correo electrónico');
-      });
-    });
-
-    describe('Mixed Dynamic Text Types', () => {
-      it('should handle different dynamic text types in same form', async () => {
-        const translationService = createTestTranslationService({
-          'form.email.label': 'Email',
-        });
-
-        const dynamicPlaceholder = new BehaviorSubject('Dynamic placeholder');
-        const staticLabel = 'Static Label';
-
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            label: translationService.translate('form.email.label'),
-          })
-          .field({
-            key: 'name',
-            type: 'input',
-            label: staticLabel,
-            placeholder: dynamicPlaceholder.asObservable(),
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '', name: '' },
-        });
-
-        const labels = fixture.debugElement.queryAll(By.css('mat-label'));
-        const inputs = fixture.debugElement.queryAll(By.css('input'));
-
-        expect(labels[0].nativeElement.textContent.trim()).toBe('Email');
-        expect(labels[1].nativeElement.textContent.trim()).toBe('Static Label');
-        expect(inputs[1].nativeElement.getAttribute('placeholder')).toBe('Dynamic placeholder');
-      });
-    });
-
-    describe('Edge Cases', () => {
-      it('should handle undefined dynamic text gracefully', async () => {
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            label: undefined,
-            placeholder: undefined,
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        const label = fixture.debugElement.query(By.css('mat-label'));
-        const input = fixture.debugElement.query(By.css('input'));
-
-        expect(label).toBeNull(); // No label rendered when undefined
-        expect(input.nativeElement.getAttribute('placeholder')).toBeNull(); // null for undefined placeholder
-      });
-
-      it('should handle empty string from Observable', async () => {
-        const emptyObservable = new BehaviorSubject('');
-
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            label: emptyObservable.asObservable(),
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        const label = fixture.debugElement.query(By.css('mat-label'));
-        expect(label.nativeElement.textContent.trim()).toBe('');
-      });
-
-      it('should handle empty Observable values', async () => {
-        const emptyObservable = of('');
-
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'email',
-            type: 'input',
-            placeholder: emptyObservable,
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { email: '' },
-        });
-
-        const input = fixture.debugElement.query(By.css('input'));
-
-        // Empty Observable results in null placeholder attribute in Material
-        expect(input.nativeElement.getAttribute('placeholder')).toBeNull();
-      });
+      expect(hostElement.nativeElement.className).toContain('email-input');
     });
   });
 });
