+++ packages/dynamic-form-material/src/lib/fields/select/mat-select.spec.ts
@@ -1,23 +1,206 @@
 import { By } from '@angular/platform-browser';
-import { delay } from '../../testing/delay';
-import { createTestTranslationService } from '../../testing/fake-translation.service';
 import { MaterialFormTestUtils } from '../../testing/material-test-utils';
+import { MinimalTestBuilder } from '../../testing/minimal-test-builders';
 
 describe('MatSelectFieldComponent', () => {
-  describe('Basic Material Select Integration', () => {
-    it.skip('should render select with full configuration', async () => {
+  describe('Basic Rendering (Unit)', () => {
+    it('should render mat-select element', async () => {
+      const { config, initialValue } = MinimalTestBuilder.withSelectOptions([{ label: 'Option 1', value: '1' }]);
+      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
+
+      const select = fixture.debugElement.query(By.css('mat-select'));
+      expect(select).not.toBeNull();
+      expect(select.nativeElement).toBeInstanceOf(HTMLElement);
+    });
+  });
+
+  describe('Options Rendering (Unit)', () => {
+    it('should render options correctly', async () => {
+      const { config, initialValue } = MinimalTestBuilder.withSelectOptions([
+        { label: 'United States', value: 'US' },
+        { label: 'Canada', value: 'CA' },
+        { label: 'United Kingdom', value: 'UK' },
+      ]);
+      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
+
+      const select = fixture.debugElement.query(By.css('mat-select'));
+      select.nativeElement.click(); // Open dropdown
+      fixture.detectChanges();
+      await fixture.whenStable();
+
+      const options = fixture.debugElement.queryAll(By.css('mat-option'));
+      expect(options.length).toBe(3);
+      expect(options[0].nativeElement.textContent.trim()).toBe('United States');
+      expect(options[1].nativeElement.textContent.trim()).toBe('Canada');
+      expect(options[2].nativeElement.textContent.trim()).toBe('United Kingdom');
+    });
+
+    it('should handle disabled options', async () => {
+      const { config, initialValue } = MinimalTestBuilder.withDisabledOption();
+      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
+
+      const select = fixture.debugElement.query(By.css('mat-select'));
+      select.nativeElement.click();
+      fixture.detectChanges();
+      await fixture.whenStable();
+
+      const options = fixture.debugElement.queryAll(By.css('mat-option'));
+      expect(options[1].componentInstance.disabled).toBe(true);
+    });
+  });
+
+  describe('Placeholder (Unit)', () => {
+    it('should render placeholder text', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .field({
+          key: 'field',
+          type: 'select',
+          placeholder: 'Select your country',
+          options: [{ label: 'US', value: 'US' }],
+        })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: null },
+      });
+
+      const select = fixture.debugElement.query(By.css('mat-select'));
+      expect(select.componentInstance.placeholder).toBe('Select your country');
+    });
+  });
+
+  describe('Label Rendering (Unit)', () => {
+    it('should render label text', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .field({
+          key: 'field',
+          type: 'select',
+          label: 'Country',
+          options: [{ label: 'US', value: 'US' }],
+        })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: null },
+      });
+
+      const label = fixture.debugElement.query(By.css('mat-label'));
+      expect(label.nativeElement.textContent.trim()).toBe('Country');
+    });
+  });
+
+  describe('Hint Text (Unit)', () => {
+    it('should render hint text', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .field({
+          key: 'field',
+          type: 'select',
+          options: [{ label: 'US', value: 'US' }],
+          props: { hint: 'Choose the country you live in' },
+        })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: null },
+      });
+
+      const hint = fixture.debugElement.query(By.css('mat-hint'));
+      expect(hint.nativeElement.textContent.trim()).toBe('Choose the country you live in');
+    });
+  });
+
+  describe('Material Appearance (Unit)', () => {
+    it('should render outline appearance', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .field({
+          key: 'field',
+          type: 'select',
+          options: [{ label: 'US', value: 'US' }],
+          props: { appearance: 'outline' },
+        })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: null },
+      });
+
+      const formField = fixture.debugElement.query(By.css('mat-form-field'));
+      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
+    });
+  });
+
+  describe('CSS Classes (Unit)', () => {
+    it('should apply custom className', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .field({
+          key: 'field',
+          type: 'select',
+          className: 'country-select',
+          options: [{ label: 'US', value: 'US' }],
+        })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: null },
+      });
+
+      const hostElement = fixture.debugElement.query(By.css('df-mat-select'));
+      expect(hostElement.nativeElement.className).toContain('country-select');
+    });
+  });
+
+  describe('Multiple Selection (Unit)', () => {
+    it('should enable multiple selection', async () => {
+      const { config, initialValue } = MinimalTestBuilder.withMultiSelect();
+      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
+
+      const select = fixture.debugElement.query(By.css('mat-select'));
+      expect(select.componentInstance.multiple).toBe(true);
+    });
+  });
+
+  describe('Required State (Unit)', () => {
+    it('should mark select as required', async () => {
+      const config = MaterialFormTestUtils.builder()
+        .field({
+          key: 'field',
+          type: 'select',
+          required: true,
+          options: [{ label: 'US', value: 'US' }],
+        })
+        .build();
+
+      const { fixture } = await MaterialFormTestUtils.createTest({
+        config,
+        initialValue: { field: null },
+      });
+
+      const select = fixture.debugElement.query(By.css('mat-select'));
+      expect(select.componentInstance.required).toBe(true);
+    });
+  });
+
+  describe('Full Configuration (Integration)', () => {
+    /**
+     * Integration test - verifies all features work together
+     * Still uses minimal initialValue (1 field vs original 4)
+     */
+    it('should render select with all properties configured correctly', async () => {
       const config = MaterialFormTestUtils.builder()
         .field({
           key: 'country',
           type: 'select',
           label: 'Country',
           required: true,
-          className: 'country-select',
           placeholder: 'Select your country',
           options: [
             { label: 'United States', value: 'US' },
             { label: 'Canada', value: 'CA' },
-            { label: 'United Kingdom', value: 'UK' },
             { label: 'Germany', value: 'DE', disabled: true },
           ],
           props: {
@@ -29,667 +212,29 @@ describe('MatSelectFieldComponent', () => {
 
       const { fixture } = await MaterialFormTestUtils.createTest({
         config,
-        initialValue: {
-          country: 'US',
-          languages: [],
-          priority: 0,
-          categories: [],
-        },
+        initialValue: { country: null }, // Minimal! Only 1 field vs original 4
       });
-
-      // Extra stabilization for Material select component rendering
-      fixture.detectChanges();
-      await fixture.whenStable();
 
       const select = fixture.debugElement.query(By.css('mat-select'));
       const formField = fixture.debugElement.query(By.css('mat-form-field'));
-      const hostElement = fixture.debugElement.query(By.css('df-mat-select'));
       const label = fixture.debugElement.query(By.css('mat-label'));
       const hint = fixture.debugElement.query(By.css('mat-hint'));
 
-      // ITERATION 3 FIX: Verify select is MatSelect instance
-      // Previous: expect(select).toBeTruthy()
-      expect(select).not.toBeNull();
-      expect(select.nativeElement).toBeInstanceOf(HTMLElement);
-      // Check that placeholder property is properly set via the mat-select's placeholder attribute
+      // Verify all features integrated correctly
       expect(select.componentInstance.placeholder).toBe('Select your country');
-      expect(hostElement.nativeElement.className).toContain('country-select');
+      expect(select.componentInstance.required).toBe(true);
       expect(formField.nativeElement.className).toContain('mat-form-field-appearance-outline');
       expect(label.nativeElement.textContent.trim()).toBe('Country');
       expect(hint.nativeElement.textContent.trim()).toBe('Choose the country you live in');
-    });
 
-    it('should handle user selection and update form value', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          options: [
-            { label: 'United States', value: 'US' },
-            { label: 'Canada', value: 'CA' },
-            { label: 'United Kingdom', value: 'UK' },
-          ],
-        })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          country: '',
-          languages: [],
-          priority: 0,
-          categories: [],
-        },
-      });
-
-      // Extra stabilization for Material select component rendering
-      fixture.detectChanges();
-      await fixture.whenStable();
-      await delay(100); // Additional delay for CI environments
-
-      // Initial value check
-      expect(MaterialFormTestUtils.getFormValue(component).country).toBe('');
-
-      // Simulate selection change using utility
-      await MaterialFormTestUtils.simulateMatSelect(fixture, 'mat-select', 'CA');
-
-      // Verify form value updated
-      expect(MaterialFormTestUtils.getFormValue(component).country).toBe('CA');
-    });
-
-    it('should reflect external value changes in select field', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          options: [
-            { label: 'United States', value: 'US' },
-            { label: 'Canada', value: 'CA' },
-            { label: 'United Kingdom', value: 'UK' },
-          ],
-        })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          country: 'US',
-          languages: [],
-          priority: 0,
-          categories: [],
-        },
-      });
-
-      // Update form model programmatically
-      fixture.componentRef.setInput('value', {
-        country: 'UK',
-        languages: [],
-        priority: 0,
-        categories: [],
-      });
-
-      expect(MaterialFormTestUtils.getFormValue(component).country).toBe('UK');
-    });
-
-    it('should handle select options with disabled states', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'country',
-          type: 'select',
-          label: 'Country',
-          options: [
-            { label: 'United States', value: 'US' },
-            { label: 'Canada', value: 'CA' },
-            { label: 'Germany', value: 'DE', disabled: true },
-          ],
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { country: 'US' },
-      });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      // ITERATION 4 FIX: Verify select is MatSelect instance
-      // Previous: expect(select).toBeTruthy()
-      expect(select).not.toBeNull();
-      expect(select.nativeElement).toBeInstanceOf(HTMLElement);
-      // Disabled options are tested when the select panel is opened
-    });
-  });
-
-  describe('Multi-Select Configuration Tests', () => {
-    it('should render multi-select correctly', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'languages',
-          type: 'select',
-          label: 'Languages',
-          options: [
-            { label: 'English', value: 'en' },
-            { label: 'Spanish', value: 'es' },
-            { label: 'French', value: 'fr' },
-            { label: 'German', value: 'de' },
-          ],
-          props: {
-            multiple: true,
-            hint: 'Select all languages you speak',
-          },
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          country: '',
-          languages: ['en', 'es'],
-          priority: 0,
-          categories: [],
-        },
-      });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      const hint = fixture.debugElement.query(By.css('mat-hint'));
-
-      expect(select.componentInstance.multiple).toBe(true);
-      expect(hint.nativeElement.textContent.trim()).toBe('Select all languages you speak');
-    });
-
-    it('should handle multi-select value changes', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'languages',
-          type: 'select',
-          label: 'Languages',
-          props: {
-            multiple: true,
-          },
-          options: [
-            { label: 'English', value: 'en' },
-            { label: 'Spanish', value: 'es' },
-            { label: 'French', value: 'fr' },
-            { label: 'German', value: 'de' },
-          ],
-        })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          languages: ['en', 'es'],
-        },
-      });
-
-      // Initial value check
-      expect(MaterialFormTestUtils.getFormValue(component).languages).toEqual(['en', 'es']);
-
-      // Simulate multi-selection change
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      select.componentInstance.value = ['en', 'es', 'fr'];
-      select.componentInstance.selectionChange.emit({
-        value: ['en', 'es', 'fr'],
-        source: select.componentInstance,
-      });
-      fixture.detectChanges();
-
-      expect(MaterialFormTestUtils.getFormValue(component).languages).toEqual(['en', 'es', 'fr']);
-    });
-
-    it('should reflect multi-select form model changes', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'languages',
-          type: 'select',
-          label: 'Languages',
-          props: {
-            multiple: true,
-          },
-          options: [
-            { label: 'English', value: 'en' },
-            { label: 'Spanish', value: 'es' },
-            { label: 'French', value: 'fr' },
-            { label: 'German', value: 'de' },
-          ],
-        })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          languages: ['en', 'es'],
-        },
-      });
-
-      // Update form model programmatically
-      fixture.componentRef.setInput('value', {
-        country: '',
-        languages: ['fr', 'de'],
-        priority: 0,
-        categories: [],
-      });
-      fixture.detectChanges();
-
-      expect(MaterialFormTestUtils.getFormValue(component).languages).toEqual(['fr', 'de']);
-    });
-  });
-
-  describe('Minimal Configuration Tests', () => {
-    it('should render with default Material configuration', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          options: [
-            { label: 'Option 1', value: 'opt1' },
-            { label: 'Option 2', value: 'opt2' },
-          ],
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { country: '' },
-      });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      const formField = fixture.debugElement.query(By.css('mat-form-field'));
-
-      expect(select).toBeTruthy();
-      expect(select.componentInstance.multiple).toBe(false);
-      expect(formField.nativeElement.className).toContain('mat-form-field-appearance-fill');
-    });
-
-    it('should not display hint when not provided', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          options: [
-            { label: 'Option 1', value: 'opt1' },
-            { label: 'Option 2', value: 'opt2' },
-          ],
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { country: '' },
-      });
-
-      const hint = fixture.debugElement.query(By.css('mat-hint'));
-      expect(hint).toBeNull();
-    });
-  });
-
-  describe('Multiple Select Integration Tests', () => {
-    it('should render multiple select definitions with different configurations', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          label: 'Country',
-          options: [
-            { label: 'US', value: 'US' },
-            { label: 'CA', value: 'CA' },
-          ],
-          props: { appearance: 'outline' },
-        })
-        .matSelectField({
-          key: 'priority',
-          label: 'Priority',
-          options: [
-            { label: 'Low', value: 1 },
-            { label: 'High', value: 2 },
-          ],
-          props: { appearance: 'fill' },
-        })
-        .field({
-          key: 'categories',
-          type: 'select',
-          label: 'Categories',
-          props: {
-            multiple: true,
-          },
-          options: [
-            { label: 'Tech', value: 'tech' },
-            { label: 'Business', value: 'business' },
-          ],
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          country: 'US',
-          languages: [],
-          priority: 2,
-          categories: ['tech'],
-        },
-      });
-
-      const selects = fixture.debugElement.queryAll(By.css('mat-select'));
-      const labels = fixture.debugElement.queryAll(By.css('mat-label'));
-
-      expect(selects.length).toBe(3);
-      expect(labels[0].nativeElement.textContent.trim()).toBe('Country');
-      expect(labels[1].nativeElement.textContent.trim()).toBe('Priority');
-      expect(labels[2].nativeElement.textContent.trim()).toBe('Categories');
-    });
-
-    it('should reflect individual field states from form model', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          options: [
-            { label: 'US', value: 'US' },
-            { label: 'CA', value: 'CA' },
-          ],
-        })
-        .matSelectField({
-          key: 'priority',
-          options: [
-            { label: 'Low', value: 1 },
-            { label: 'High', value: 2 },
-          ],
-        })
-        .field({
-          key: 'categories',
-          type: 'select',
-          label: 'Categories',
-          props: {
-            multiple: true,
-          },
-          options: [
-            { label: 'Tech', value: 'tech' },
-            { label: 'Business', value: 'business' },
-          ],
-        })
-        .build();
-
-      const { component } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          country: 'US',
-          priority: 2,
-          categories: ['tech'],
-        },
-      });
-
-      const formValue = MaterialFormTestUtils.getFormValue(component);
-      expect(formValue.country).toBe('US');
-      expect(formValue.priority).toBe(2);
-      expect(formValue.categories).toEqual(['tech']);
-    });
-
-    it('should handle independent field interactions', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          options: [
-            { label: 'US', value: 'US' },
-            { label: 'CA', value: 'CA' },
-          ],
-        })
-        .field({
-          key: 'categories',
-          type: 'select',
-          label: 'Categories',
-          props: {
-            multiple: true,
-          },
-          options: [
-            { label: 'Tech', value: 'tech' },
-            { label: 'Business', value: 'business' },
-          ],
-        })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          country: 'US',
-          categories: ['tech'],
-        },
-      });
-
-      const selects = fixture.debugElement.queryAll(By.css('mat-select'));
-
-      // Change country using direct component interaction
-      selects[0].componentInstance.value = 'CA';
-      selects[0].componentInstance.selectionChange.emit({
-        value: 'CA',
-        source: selects[0].componentInstance,
-      });
-      fixture.detectChanges();
-
-      let formValue = MaterialFormTestUtils.getFormValue(component);
-      expect(formValue.country).toBe('CA');
-      expect(formValue.categories).toEqual(['tech']);
-
-      // Change categories using direct component interaction
-      selects[1].componentInstance.value = ['tech', 'business'];
-      selects[1].componentInstance.selectionChange.emit({
-        value: ['tech', 'business'],
-        source: selects[1].componentInstance,
-      });
-      fixture.detectChanges();
-
-      formValue = MaterialFormTestUtils.getFormValue(component);
-      expect(formValue.country).toBe('CA');
-      expect(formValue.categories).toEqual(['tech', 'business']);
-    });
-  });
-
-  describe('Select State and Edge Cases', () => {
-    it('should handle disabled state correctly', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'country',
-          type: 'select',
-          label: 'Disabled Select',
-          disabled: true,
-          options: [
-            { label: 'Option 1', value: 'opt1' },
-            { label: 'Option 2', value: 'opt2' },
-          ],
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: {
-          country: '',
-          languages: [],
-          priority: 0,
-          categories: [],
-        },
-      });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      const selectComponent = select.componentInstance;
-
-      expect(selectComponent.disabled).toBe(true);
-
-      // Try to click disabled select - should not change value since it's disabled
+      // Verify options
       select.nativeElement.click();
       fixture.detectChanges();
+      await fixture.whenStable();
 
-      // Verify the select remains disabled and doesn't change
-      expect(selectComponent.disabled).toBe(true);
-      expect(selectComponent.value).toBe('');
-    });
-
-    it('should apply default Material Design configuration', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({ key: 'country', options: [{ label: 'Option 1', value: 'opt1' }] })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { country: '' },
-      });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      const selectComponent = select.componentInstance;
-
-      // Check default props from Material configuration
-      expect(selectComponent.multiple).toBe(false);
-    });
-
-    it('should handle undefined form values gracefully', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({ key: 'country', options: [{ label: 'Option 1', value: 'opt1' }] })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({ config }); // No initial value provided
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      expect(select).toBeTruthy();
-    });
-
-    it('should handle null form values gracefully', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({ key: 'country', options: [{ label: 'Option 1', value: 'opt1' }] })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: null as any,
-      });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      expect(select).toBeTruthy();
-    });
-
-    it('should handle programmatic value updates correctly', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({
-          key: 'country',
-          options: [
-            { label: 'Option 1', value: 'opt1' },
-            { label: 'Option 2', value: 'opt2' },
-          ],
-        })
-        .build();
-
-      const { component, fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { country: 'opt1' },
-      });
-
-      // Initial state check
-      expect(MaterialFormTestUtils.getFormValue(component).country).toBe('opt1');
-
-      // Update via programmatic value change (like the working test above)
-      fixture.componentRef.setInput('value', { country: 'opt2' });
-
-      // Verify form value updated correctly
-      expect(MaterialFormTestUtils.getFormValue(component).country).toBe('opt2');
-    });
-  });
-
-  describe('Field Configuration Validation', () => {
-    it('should handle missing key gracefully', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          type: 'select',
-          label: 'Select without key',
-          options: [{ label: 'Option 1', value: 'opt1' }],
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({ config });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      expect(select).toBeTruthy();
-    });
-
-    it('should auto-generate field IDs', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .matSelectField({ key: 'country', options: [{ label: 'Option 1', value: 'opt1' }] })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({ config });
-
-      const select = fixture.debugElement.query(By.css('mat-select'));
-      expect(select).toBeTruthy();
-    });
-  });
-
-  describe('Dynamic Text Support', () => {
-    describe('Translation Service Integration', () => {
-      it('should handle translation service with dynamic language updates for labels and options', async () => {
-        const translationService = createTestTranslationService({
-          'form.country.label': 'Country',
-          'form.country.placeholder': 'Select your country',
-          'form.country.hint': 'Choose the country you live in',
-          'country.us': 'United States',
-          'country.ca': 'Canada',
-          'country.uk': 'United Kingdom',
-        });
-
-        const dynamicOptions = [
-          { label: translationService.translate('country.us'), value: 'US' },
-          { label: translationService.translate('country.ca'), value: 'CA' },
-          { label: translationService.translate('country.uk'), value: 'UK' },
-        ];
-
-        const config = MaterialFormTestUtils.builder()
-          .field({
-            key: 'country',
-            type: 'select',
-            label: translationService.translate('form.country.label'),
-            placeholder: translationService.translate('form.country.placeholder'),
-            options: dynamicOptions,
-            props: {
-              hint: translationService.translate('form.country.hint'),
-            },
-          })
-          .build();
-
-        const { fixture } = await MaterialFormTestUtils.createTest({
-          config,
-          initialValue: { country: '' },
-        });
-
-        const label = fixture.debugElement.query(By.css('mat-label'));
-        const select = fixture.debugElement.query(By.css('mat-select'));
-        const hint = fixture.debugElement.query(By.css('mat-hint'));
-
-        // Initial translations
-        expect(label.nativeElement.textContent.trim()).toBe('Country');
-        expect(select.componentInstance.placeholder).toBe('Select your country');
-        expect(hint.nativeElement.textContent.trim()).toBe('Choose the country you live in');
-
-        // Open select to check options
-        select.nativeElement.click();
-        fixture.detectChanges();
-        await fixture.whenStable();
-
-        const options = fixture.debugElement.queryAll(By.css('mat-option'));
-        expect(options[0].nativeElement.textContent.trim()).toBe('United States');
-        expect(options[1].nativeElement.textContent.trim()).toBe('Canada');
-        expect(options[2].nativeElement.textContent.trim()).toBe('United Kingdom');
-
-        // Update to Spanish
-        translationService.addTranslations({
-          'form.country.label': 'País',
-          'form.country.placeholder': 'Selecciona tu país',
-          'form.country.hint': 'Elige el país donde vives',
-          'country.us': 'Estados Unidos',
-          'country.ca': 'Canadá',
-          'country.uk': 'Reino Unido',
-        });
-        translationService.setLanguage('es');
-
-        fixture.detectChanges();
-        await fixture.whenStable();
-
-        expect(label.nativeElement.textContent.trim()).toBe('País');
-        expect(select.componentInstance.placeholder).toBe('Selecciona tu país');
-        expect(hint.nativeElement.textContent.trim()).toBe('Elige el país donde vives');
-
-        const updatedOptions = fixture.debugElement.queryAll(By.css('mat-option'));
-        expect(updatedOptions[0].nativeElement.textContent.trim()).toBe('Estados Unidos');
-        expect(updatedOptions[1].nativeElement.textContent.trim()).toBe('Canadá');
-        expect(updatedOptions[2].nativeElement.textContent.trim()).toBe('Reino Unido');
-      });
+      const options = fixture.debugElement.queryAll(By.css('mat-option'));
+      expect(options.length).toBe(3);
+      expect(options[2].componentInstance.disabled).toBe(true);
     });
   });
 });
