--- packages/dynamic-form-material/src/lib/fields/slider/mat-slider-refactored.spec.ts
@@ -1,241 +1,0 @@
-import { By } from '@angular/platform-browser';
-import { MaterialFormTestUtils } from '../../testing/material-test-utils';
-import { MinimalTestBuilder } from '../../testing/minimal-test-builders';
-
-describe('MatSliderFieldComponent (Refactored)', () => {
-  describe('Basic Rendering (Unit)', () => {
-    it('should render slider element', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderRange(0, 100);
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      expect(slider).not.toBeNull();
-      expect(slider.nativeElement).toBeInstanceOf(HTMLElement);
-    });
-  });
-
-  describe('Min/Max Range (Unit)', () => {
-    it('should apply min and max values', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderRange(0, 100);
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      expect(slider.componentInstance.min).toBe(0);
-      expect(slider.componentInstance.max).toBe(100);
-    });
-
-    it('should apply different min/max range', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderRange(-50, 50);
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      expect(slider.componentInstance.min).toBe(-50);
-      expect(slider.componentInstance.max).toBe(50);
-    });
-  });
-
-  describe('Step Value (Unit)', () => {
-    it('should apply step value', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderStep(5);
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const sliderInput = fixture.debugElement.query(By.css('input[matSliderThumb]'));
-      expect(sliderInput.nativeElement.getAttribute('step')).toBe('5');
-    });
-  });
-
-  describe('Label Rendering (Unit)', () => {
-    it('should render label text', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({ key: 'field', type: 'slider', label: 'Volume Level', props: { min: 0, max: 100 } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: 0 },
-      });
-
-      const label = fixture.debugElement.query(By.css('.slider-label'));
-      expect(label.nativeElement.textContent.trim()).toBe('Volume Level');
-    });
-
-    it('should not render label when not provided', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderRange(0, 100);
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const label = fixture.debugElement.query(By.css('.slider-label'));
-      expect(label).toBeNull();
-    });
-  });
-
-  describe('Hint Text (Unit)', () => {
-    it('should render hint text', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({ key: 'field', type: 'slider', props: { min: 0, max: 100, hint: 'Adjust the volume level' } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: 0 },
-      });
-
-      const hint = fixture.debugElement.query(By.css('.mat-hint'));
-      expect(hint.nativeElement.textContent.trim()).toBe('Adjust the volume level');
-    });
-
-    it('should not render hint when not provided', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderRange(0, 100);
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const hint = fixture.debugElement.query(By.css('.mat-hint'));
-      expect(hint).toBeNull();
-    });
-  });
-
-  describe('CSS Classes (Unit)', () => {
-    it('should apply custom className', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({ key: 'field', type: 'slider', className: 'volume-slider', props: { min: 0, max: 100 } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: 0 },
-      });
-
-      const container = fixture.debugElement.query(By.css('.volume-slider'));
-      expect(container).not.toBeNull();
-      expect(container.nativeElement.classList.contains('volume-slider')).toBe(true);
-    });
-  });
-
-  describe('Material Color Theme (Unit)', () => {
-    it('should apply primary color', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderColor('primary');
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      expect(slider.componentInstance.color).toBe('primary');
-    });
-
-    it('should apply accent color', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderColor('accent');
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      expect(slider.componentInstance.color).toBe('accent');
-    });
-  });
-
-  describe('Thumb Label (Unit)', () => {
-    it('should enable thumb label', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderThumbLabel();
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      expect(slider.componentInstance.discrete).toBe(true);
-    });
-  });
-
-  describe('Tick Marks (Unit)', () => {
-    it('should display tick marks', async () => {
-      const { config, initialValue } = MinimalTestBuilder.withSliderTickMarks(25);
-      const { fixture } = await MaterialFormTestUtils.createTest({ config, initialValue });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      expect(slider.componentInstance.showTickMarks).toBe(true);
-    });
-  });
-
-  describe('TabIndex (Unit)', () => {
-    it('should apply tabindex attribute', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({ key: 'field', type: 'slider', tabIndex: 5, props: { min: 0, max: 100 } })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { field: 0 },
-      });
-
-      const sliderInput = fixture.debugElement.query(By.css('input[matSliderThumb]'));
-      expect(sliderInput.nativeElement.getAttribute('tabindex')).toBe('5');
-    });
-  });
-
-  describe('Full Configuration (Integration)', () => {
-    it('should render slider with all properties configured correctly', async () => {
-      const config = MaterialFormTestUtils.builder()
-        .field({
-          key: 'volume',
-          type: 'slider',
-          label: 'Volume Level',
-          tabIndex: 1,
-          className: 'volume-slider',
-          props: {
-            min: 0,
-            max: 100,
-            step: 5,
-            thumbLabel: true,
-            showThumbLabel: true,
-            hint: 'Adjust the volume level',
-            color: 'primary',
-            tickInterval: 25,
-          },
-        })
-        .build();
-
-      const { fixture } = await MaterialFormTestUtils.createTest({
-        config,
-        initialValue: { volume: 50 }, // Minimal! Only 1 field vs original 5
-      });
-
-      const slider = fixture.debugElement.query(By.css('mat-slider'));
-      const sliderInput = fixture.debugElement.query(By.css('input[matSliderThumb]'));
-      const label = fixture.debugElement.query(By.css('.slider-label'));
-      const hint = fixture.debugElement.query(By.css('.mat-hint'));
-      const container = fixture.debugElement.query(By.css('.volume-slider'));
-
-      // Verify all features integrated correctly
-      expect(slider).not.toBeNull();
-      expect(slider.nativeElement).toBeInstanceOf(HTMLElement);
-      expect(slider.componentInstance.min).toBe(0);
-      expect(slider.componentInstance.max).toBe(100);
-      expect(slider.componentInstance.discrete).toBe(true);
-      expect(slider.componentInstance.showTickMarks).toBe(true);
-      expect(slider.componentInstance.color).toBe('primary');
-      expect(sliderInput.nativeElement.getAttribute('tabindex')).toBe('1');
-      expect(container).not.toBeNull();
-      expect(container.nativeElement.classList.contains('volume-slider')).toBe(true);
-      expect(label.nativeElement.textContent.trim()).toBe('Volume Level');
-      expect(hint.nativeElement.textContent.trim()).toBe('Adjust the volume level');
-    });
-  });
-});
-
-/**
- * IMPACT ANALYSIS:
- *
- * Original (mat-slider.spec.ts:8-65):
- * - 1 test checking 8 properties
- * - initialValue: { volume: 50, brightness: 0, rating: 0, temperature: 0, speed: 0 }
- * - Runtime: 909ms (3RD SLOWEST!)
- * - Failure: Generic error messages
- *
- * Refactored (this file):
- * - 15 focused tests (13 unit + 2 integration)
- * - initialValue: { field: 0 } for each test
- * - Estimated runtime: ~1700ms total (800ms first + 14 tests @ ~100ms avg)
- * - Failure: Specific error messages
- *
- * NET CHANGE:
- * - Slightly slower overall (~791ms more = +87%)
- * - BUT: 15x more granular, 100% clearer failures
- * - Bug discovery: TBD (need to check component implementation)
- * - Parallelizable: YES (can run 4 tests concurrently)
- *
- * WITH PARALLELIZATION (4 cores):
- * - Sequential: 1700ms
- * - Parallel: ~1000ms (800 + ceiling(14/4) * 100)
- * - Net: 10% faster than original!
- */
